@@ -17,69 +17,27 @@ limitations under the License.
 package main
 
 import (
-	"context"
-	"fmt"
-	"os"
-	"strings"
-
-	"github.com/prometheus/common/version"
-	"gopkg.in/yaml.v3"
+	"github.com/spf13/cobra"
 	"k8s.io/klog/v2"
 
-	"k8s.io/kube-state-metrics/v2/pkg/customresource"
-	"k8s.io/kube-state-metrics/v2/pkg/customresourcestate"
-
-	"k8s.io/kube-state-metrics/v2/pkg/app"
+	"k8s.io/kube-state-metrics/v2/internal"
 	"k8s.io/kube-state-metrics/v2/pkg/options"
 )
 
 func main() {
 	opts := options.NewOptions()
-	opts.AddFlags()
+	cmd := options.InitCommand
+	cmd.Run = func(cmd *cobra.Command, args []string) {
+		internal.RunKubeStateMetricsWrapper(opts)
+	}
+	opts.AddFlags(cmd)
 
 	if err := opts.Parse(); err != nil {
-		klog.ErrorS(err, "Parsing flag definitions error")
 		klog.FlushAndExit(klog.ExitFlushTimeout, 1)
 	}
 
-	if opts.Version {
-		fmt.Printf("%s\n", version.Print("kube-state-metrics"))
-		os.Exit(0)
-	}
-
-	if opts.Help {
-		opts.Usage()
-		os.Exit(0)
-	}
-
-	var factories []customresource.RegistryFactory
-	if config, set := resolveCustomResourceConfig(opts); set {
-		crf, err := customresourcestate.FromConfig(config)
-		if err != nil {
-			klog.ErrorS(err, "Parsing from Custom Resource State Metrics file failed")
-			klog.FlushAndExit(klog.ExitFlushTimeout, 1)
-		}
-		factories = append(factories, crf...)
-	}
-
-	ctx := context.Background()
-	if err := app.RunKubeStateMetrics(ctx, opts, factories...); err != nil {
-		klog.ErrorS(err, "Failed to run kube-state-metrics")
+	if err := opts.Validate(); err != nil {
+		klog.ErrorS(err, "Validating options error")
 		klog.FlushAndExit(klog.ExitFlushTimeout, 1)
 	}
-}
-
-func resolveCustomResourceConfig(opts *options.Options) (customresourcestate.ConfigDecoder, bool) {
-	if s := opts.CustomResourceConfig; s != "" {
-		return yaml.NewDecoder(strings.NewReader(s)), true
-	}
-	if file := opts.CustomResourceConfigFile; file != "" {
-		f, err := os.Open(file)
-		if err != nil {
-			klog.ErrorS(err, "Custom Resource State Metrics file could not be opened")
-			klog.FlushAndExit(klog.ExitFlushTimeout, 1)
-		}
-		return yaml.NewDecoder(f), true
-	}
-	return nil, false
 }
