@@ -16,28 +16,28 @@ limitations under the License.
 
 package metricsstore
 
-import "io"
+import (
+	"fmt"
+	"io"
+)
 
-// MetricsWriter is the interface that wraps the WriteAll method.
-// WriteAll writes out bytes to the underlying writer.
-type MetricsWriter interface {
-	WriteAll(w io.Writer)
-}
+// MetricsWriterList represent a list of MetricsWriter
+type MetricsWriterList []*MetricsWriter
 
-// MultiStoreMetricsWriter is a struct that holds multiple MetricsStore(s) and
+// MetricsWriter is a struct that holds multiple MetricsStore(s) and
 // implements the MetricsWriter interface.
 // It should be used with stores which have the same metric headers.
 //
-// MultiStoreMetricsWriter writes out metrics from the underlying stores so that
+// MetricsWriter writes out metrics from the underlying stores so that
 // metrics with the same name coming from different stores end up grouped together.
 // It also ensures that the metric headers are only written out once.
-type MultiStoreMetricsWriter struct {
+type MetricsWriter struct {
 	stores []*MetricsStore
 }
 
-// NewMultiStoreMetricsWriter creates a new MultiStoreMetricsWriter.
-func NewMultiStoreMetricsWriter(stores []*MetricsStore) MetricsWriter {
-	return &MultiStoreMetricsWriter{
+// NewMetricsWriter creates a new MetricsWriter.
+func NewMetricsWriter(stores ...*MetricsStore) *MetricsWriter {
+	return &MetricsWriter{
 		stores: stores,
 	}
 }
@@ -46,9 +46,9 @@ func NewMultiStoreMetricsWriter(stores []*MetricsStore) MetricsWriter {
 //
 // WriteAll writes metrics so that the ones with the same name
 // are grouped together when written out.
-func (m MultiStoreMetricsWriter) WriteAll(w io.Writer) {
+func (m MetricsWriter) WriteAll(w io.Writer) error {
 	if len(m.stores) == 0 {
-		return
+		return nil
 	}
 
 	for _, s := range m.stores {
@@ -59,12 +59,19 @@ func (m MultiStoreMetricsWriter) WriteAll(w io.Writer) {
 	}
 
 	for i, help := range m.stores[0].headers {
-		w.Write([]byte(help))
-		w.Write([]byte{'\n'})
+		_, err := w.Write([]byte(help + "\n"))
+		if err != nil {
+			return fmt.Errorf("failed to write help text: %v", err)
+		}
+
 		for _, s := range m.stores {
 			for _, metricFamilies := range s.metrics {
-				w.Write(metricFamilies[i])
+				_, err := w.Write(metricFamilies[i])
+				if err != nil {
+					return fmt.Errorf("failed to write metrics family: %v", err)
+				}
 			}
 		}
 	}
+	return nil
 }
