@@ -55,7 +55,11 @@ func TestAsLibrary(t *testing.T) {
 	time.Sleep(time.Second)
 
 	w := strings.Builder{}
-	c.WriteAll(&w)
+	mw := metricsstore.NewMetricsWriter(c)
+	err = mw.WriteAll(&w)
+	if err != nil {
+		t.Fatalf("failed to write metrics: %v", err)
+	}
 	m := w.String()
 
 	if !strings.Contains(m, service.ObjectMeta.Name) {
